userNo 받아서 스티커, 보드 테이블에 저장하기

+++ src/main/java/com/ilovepc/project_home/web/sticker/vo/BoardParam.java
@@ -7,7 +7,7 @@ import lombok.Builder;
 @HomeMaster
 public class BoardParam {
     private Long   id;    //생성된 key 전달받기용임!
-    private Long   userNo;         // '사용자 번호'
+    private int   userNo;         // '사용자 번호'
     private String title;          //'보드판 제목'
     private String goal;           // '칭찬 목표'
     private int    totalSlots;     // '전체 칸 수'
